+++ R/plotSpectra.R
@@ -35,12 +35,12 @@
 #' The plotting of the legend is based on the origin of the plot as [0,0] and the top right
 #' corner is [1,1]. If one wants the plot dead center use \code{leg.loc = list(x = 0.5, y = 0.5)}.
 #'
-#' @param \dots Additional parameters to be passed to plotting functions.  Names will depend
-#'        upon the choice of graphics options.
+#' @param \dots Parameters to be passed to the plotting routines. Arguments must be consistent with the 
+#'        choice of graphics option. See \code{\link{GraphicsOptions}}.
 #'
 #' @return
 #' The returned value depends on the graphics option selected (see \code{\link{GraphicsOptions}}).
-#' \itemize{
+#' \describe{
 #'  \item{base:}{    None.  Side effect is a plot.}
 #'  \item{ggplot2:}{    Returns a \code{ggplot2} plot object. The plot can be modified in the usual
 #'                      \code{ggplot2} manner.}
@@ -55,43 +55,6 @@
 #'
 #' @keywords hplot
 #'
-#' @examples
-#'
-#' data(metMUD1)
-#'
-#' # Using base graphics (the default)
-#' plotSpectra(metMUD1,
-#'   main = "metMUD1 NMR Data",
-#'   which = c(10, 11), yrange = c(0, 1.5),
-#'   offset = 0.06, amplify = 10, lab.pos = 0.5
-#' )
-#'
-#' # Add a legend at x, y coords
-#' plotSpectra(metMUD1,
-#'   main = "metMUD1 NMR Data",
-#'   which = c(10, 11), yrange = c(0, 1.5),
-#'   offset = 0.06, amplify = 10, lab.pos = 0.5,
-#'   leg.loc = list(x = 0.8, y = 0.8)
-#' )
-#'
-#' # Using ggplot2 graphics
-#' options(ChemoSpecGraphics = "ggplot2")
-#' p <- plotSpectra(metMUD1,
-#'   main = "metMUD1 NMR Data",
-#'   which = c(10, 11), yrange = c(0, 1.5),
-#'   offset = 0.06, amplify = 10, lab.pos = 0.5
-#' )
-#' p
-#'
-#' # Add a legend at x, y coords
-#' p <- plotSpectra(metMUD1,
-#'   main = "metMUD1 NMR Data",
-#'   which = c(10, 11), yrange = c(0, 1.5),
-#'   offset = 0.06, amplify = 10, lab.pos = 0.5,
-#'   leg.loc = list(x = 0.8, y = 0.8)
-#' )
-#' p
-#'
 #' @export plotSpectra
 #'
 #' @importFrom graphics grid lines text points plot
@@ -100,6 +63,29 @@
 #' @importFrom ggplot2 scale_color_manual theme theme_bw theme_classic ylim
 #' @importFrom grid grobTree textGrob
 #' @importFrom reshape2 melt
+#' @examples
+#'
+#' data(metMUD1)
+#'
+#' # Using base graphics (the default, but need to set here
+#' # in case someone changed it!)
+#' options(ChemoSpecGraphics = "ggplot2")
+#' plotSpectra(metMUD1,
+#'   main = "metMUD1 NMR Data",
+#'   which = c(10, 11), yrange = c(0, 1.5),
+#'   offset = 0.06, amplify = 10, lab.pos = 0.5
+#' )
+#'
+#' # Add a legend at x, y coords
+#' plotSpectra(metMUD1,
+#'   main = "metMUD1 NMR Data",
+#'   which = c(10, 11), yrange = c(0, 1.5),
+#'   offset = 0.06, amplify = 10, lab.pos = 0.5,
+#'   leg.loc = list(x = 0.8, y = 0.8)
+#' )
+#'
+#' options(ChemoSpecGraphics = "base") # turn off for later examples/tests
+#'
 #'
 
 plotSpectra <- function(spectra, which = c(1),
@@ -110,38 +96,52 @@ plotSpectra <- function(spectra, which = c(1),
   .chkArgs(mode = 11L)
   chkSpectra(spectra)
 
-  go <- chkGraphicsOpt()
+  # Helper Function to calculate the label y position
+  # spec: spectra data matrix with *samples in rows* (previously subsetted by which,
+  #       & modified by offset & amplify)
+  # pct: label position in % of y range
+  pct = 20.0
+  calcYpos <- function(spec, pct) {
+    ymin <- apply(spec, 1, min)
+    ymax <- apply(spec, 1, max)
+    ypos <- ymin + (pct * (ymax - ymin))/100
+  }
 
+  go <- chkGraphicsOpt()
 
   if (go == "base") {
 
-    # set up and plot the first spectrum (type = "n")
+    # Prepare the data needed for plotting, apply amplify & offset
+    M <- spectra$data[which,, drop = FALSE]
+    Mcols <- spectra$colors[which]
+    Mnames <- spectra$names[which]
+    freq <- spectra$freq
+    count <- 0L
+    for (i in 1:nrow(M)) {
+      M[i, ] <- (M[i,] + (offset * count)) * amplify
+      count <- count + 1
+    }
 
-    spectrum <- spectra$data[which[1], ] * amplify
-    plot(spectra$freq, spectrum,
+    # Set up and plot the first spectrum (type = "n")
+    plot(freq, M[1,],
       type = "n",
       xlab = spectra$unit[1], ylab = spectra$unit[2],
       ylim = yrange,
       frame.plot = FALSE, ...
     )
 
-    if (showGrid) grid(ny = NA, lty = 1) # grid will be underneath all spectra
-    lines(spectra$freq, spectrum, col = spectra$colors[which[1]], ...)
+    # Place grid under all spectra
+    if (showGrid) grid(ny = NA, lty = 1)
+
+    # Add the spectra
+    for (i in 1:nrow(M)) lines(freq, M[i, ], col = Mcols[i], ...)
+
+    # Add sample names
     lab.x <- lab.pos
-    freq.index <- findInterval(lab.x, sort(spectra$freq))
-    lab.y <- spectrum[freq.index]
-    text(lab.x, lab.y, labels = spectra$names[which[1]], pos = 3, cex = 0.75)
+    lab.y <- calcYpos(M, pct)
+    text(lab.x, lab.y, labels = Mnames, adj = c(0.5, 0.5), cex = 0.75)
 
-    which <- which[-1] # first spectrum already plotted so remove it from the list
-    count <- 0 # get the other spectra and plot them as well
-    for (n in which) {
-      count <- count + 1
-      spectrum <- (spectra$data[n, ] + (offset * count)) * amplify
-      points(spectra$freq, spectrum, type = "l", col = spectra$colors[n], ...)
-      lab.y <- spectrum[freq.index]
-      text(lab.x, lab.y, labels = spectra$names[n], pos = 3, cex = 0.75)
-    }
-
+    # Prep legend location if legend requested
     if (all(leg.loc != "none")) {
       x.min <- min(spectra$freq)
       x.max <- max(spectra$freq)
@@ -167,10 +167,9 @@ plotSpectra <- function(spectra, which = c(1),
   if (go == "ggplot2") {
     value <- variable <- Frequency <- NULL # satisfy CRAN check engine
 
-    # Set up data frame for plotting
+    # Set up data frame to hold data to be plotted ready for melting
     df <- data.frame(spectra$freq)
     count <- 0
-
     for (i in which) {
       spec <- ((spectra$data[i, ]) + (count * offset)) * amplify
       df <- cbind(df, spec)
@@ -178,18 +177,8 @@ plotSpectra <- function(spectra, which = c(1),
     }
     names(df) <- c("Frequency", spectra$names[which])
 
-    lab.x <- lab.pos
-    lab.y <- c(NA_real_)
-
-    for (i in 2:ncol(df)) {
-      spec.max <- max(df[, i])
-      spec.min <- min(df[, i])
-      # put the label at 30 % of the total height for each spectrum
-      pos.y <- spec.min + (30 * (spec.max - spec.min)) / 100
-      lab.y <- c(lab.y, pos.y)
-    }
-
-    lab.y <- lab.y[-1] # remove the first value as it is NA_real_
+    lab.x <- lab.pos # values in native data space
+    lab.y <- calcYpos(t(as.matrix(df[,-1])), pct)
 
     molten.data <- reshape2::melt(df, id = c("Frequency"))
 
@@ -200,7 +189,6 @@ plotSpectra <- function(spectra, which = c(1),
       annotate("text", x = lab.x, y = lab.y, label = spectra$names[which]) +
       labs(x = spectra$unit[1], y = spectra$unit[2]) +
       ylim(yrange) +
-      # theme_classic() +
       theme_bw() +
       theme(legend.position = "none") +
       theme(panel.border = element_blank(), axis.line = element_line(colour = "black")) +
@@ -232,32 +220,19 @@ plotSpectra <- function(spectra, which = c(1),
 
       keys <- grobTree(textGrob("Key",
         x = leg.loc$x, y = leg.loc$y + 0.025, hjust = 0,
-        gp = gpar(col = "black", fontsize = 12, fontface = "italic")
+        gp = gpar(col = "black", fontsize = 12)
       ))
 
       for (i in 1:length(group))
       {
         grob <- grid::grobTree(textGrob(group[i],
           x = leg.loc$x, y = leg.loc$y, hjust = 0,
-          gp = gpar(col = color[i], fontsize = 12, fontface = "italic")
+          gp = gpar(col = color[i], fontsize = 12)
         ))
         leg.loc$y <- leg.loc$y - 0.025
         p <- p + annotation_custom(grob) + annotation_custom(keys)
       }
     }
-    # args <- as.list(match.call())[-1] # Capturing the xlim
-
-    # if ("xlim" %in% names(args)) {
-      # xl <- eval(args$xlim)
-      # p <- p + coord_cartesian(xlim = c(xl[1], xl[2])) # Zooming in the plot according to xlim range
-    # }
-
-    # if ("main" %in% names(args)) # Capturing main
-      # {
-        # yl <- eval(args$main)
-        # p <- p + ggtitle(yl[1]) # Title of the plot
-        # p <- p + theme(plot.title = element_text(hjust = 0.5)) # Aligning the title to center
-      # }
 
     return(p)
   }